ENG-4232: Updated C++ Cassandra driver version to 2.9.0-yb-2. (Fixed RapidJson version.)
Summary:
Changed C++ Cassandra driver version in the driver thirdparty script.
In the version 2.9.0-yb-2 version of the RapidJson lib is sync-ed with the lib version in our thirdparty folder.
It prevents code crash in the RapidJson lib if the driver lib is linked.

Test Plan: ybd --cxx-test integration-tests_cassandra_cpp_driver-test

Reviewers: timur

Reviewed By: timur

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D5742