ENG-2987: #1278: Use libc++ with clang in linux
Summary:
Currently when we are building debug or release with clang,
we use libstdc++ and 3rd party libraries built by GCC.

But libstdc++ is not well compatible with clang, for instance
it fails to compile `std::atomic<std::chrono::steady_clock::time_point>`.

Also, there could be problems when using libraries built with GCC.
For instance, libcds crashes in this case.

Added new build type to thirdparty, and updated build process to use libc++ for clang.

Test Plan: Jenkins

Reviewers: mikhail, timur

Reviewed By: timur

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D6535