Add flex, bison, libuuid to the third-party build system (#14)
This diff helps reduce our dependencies on Homebrew/Linuxbrew.  We are
adding flex, bison, and libuuid to thirdparty and using a Linuxbrew
tarball that does not have the respective packages and does not have
Ninja either. Ninja is supposed to be installed on the build host
outside of our third-party build system.

Other improvements:
- Support running autoreconf in build_with_configure. Needed when the
  configure script is not present.
- Extract archives in a temporary directory, check if there is only one
  resulting directory, and rename that directory to the specified name.
  This removes any assumptions about what the directory inside the
  archive is called.