Support Clang 10, 11 on CentOS 7 and Ubuntu 18.04, GCC 8, 9 devtoolset compilers on CentOS, major refactoring (#44)
Supporting additional compilers:
- Clang 10 and Clang 11 compilers on CentOS 7 and Ubuntu 18.04.
- GCC 8 and GCC 9 (devtoolset-based) on CentOS 7, including initial support for ASAN.

These builds do not use Linuxbrew and do not build LLVM. This reduces the compilation time considerably. We will now be producing many more third-party archives and we will need to select the right one during the YugabyteDB build. The changes to the yugabyte/yugabyte-db still need to be made separately.

Moving the icu4c dependency to the "instrumented" section as it contains C++ code. Previously, there was a possibility that we were using libstdc++ for icu4c, and using libc++ for the rest of the code in Clang based builds (including ASAN/TSAN).

Refactoring / cleanup:
- Add a ./check_python_code.sh script that checks Python code using MyPy, syntax check, doctest, unit test, etc.
- Split the code into multiple files.
- Got rid of the old download URL for the few remaining dependencies (yugabyte/yugabyte-db#4747), moving all downloads to GitHub, and adding a fallback URL.
- Add a "multi build" option for testing. It is similar to what we do using CircleCI, but allows for interactive debugging in the resulting Docker containers. This is still an experimental option.