Do not use malloc/free in libunwind to maintain async safety (#54)
We collect stack traces from signal handlers, and malloc/free are not allowed in signal handlers. Specify _LIBUNWIND_NO_HEAP when building libunwind that ships with LLVM 10+. This does not affect how we build libunwind for GCC build types or the Clang 7 build.

Original LLVM diff: https://reviews.llvm.org/D11897

Also renaming the zlib.py file to zlib_dependency.py to avoid a conflict with the built-in zlib module in Python.