Enable Clang 14 and update LLVM URLs (#107)
Updating dependencies to use our custom forks:
- zlib (contains patches to correctly set CC and allow customizing TEST_LDFLAGS).
- protobuf (remove JavaScript generator because we don't use it and it breaks compilation with LTO on Clang 14).

Clang 14 has a bug causing ASAN_OPTIONS=detect_odr_violation=0 to not work, resulting in icu4c build failing under ASAN. To work around this, we are changing the corresponding step to copy the files it is trying to generate from the uninstrumented build of icu4c.

Adding a lot of CentOS 7 build types (non-LTO, thin and full LTO) because CentOS 7 builds might be valuable as the candidates for the new "universal build" depending on an old enough glibc version so that it would work on most Linux distributions. We might cut down the set of these combinations later.

Updating LLVM tarball URLs, adding Amazon Linux 2 LLVM packages, as well as Clang 13 and Clang 14 across various OS versions.