[yugabyte/yugabyte-db#12089] Add ncurses dependency; upgrade to OpenSSL 1.1.1n (#108)
Issue: https://github.com/yugabyte/yugabyte-db/issues/12089

This is one of the few remaining dependencies we have on system packages. It is because of ncurses and glibc that we can't take code built for CentOS 7 and run it e.g. on RHEL 8. If we build all of dependencies ourselves, we will be able to produce a "universal Linux" third-party library archive for each architecture.

Found that we were duplicating include flags in the CPPFLAGS (meaning "C preprocessor flags") environment variable as well as in CXXFLAGS and CFLAGS variables, which causes failures to build with LLVM with custom glibc (i.e. while using Linuxbrew). Refactored the framework so that each dependency class can specify whether to use the CPPFLAGS variable by overriding the `use_cppflags_env_var` method.

Also upgrading to OpenSSL 1.1.1n, and cleaning up the code by renaming "llvm1x" to simply "llvm" (we don't support LLVM versions less than 10 anymore).