Use C++20 for build on macOS when possible (#139)
Previously, we used C++17 for building C++ code on macOS. This created certain problems with building YugabyteDB with precompiled headers.

In this commit, we are modifying the build to use C++2a on macOS for Clang 12, and C++20 for Clang 13 or later. For the icu4c library, as a special case, we build it with C++17 when using Clang 12 or earlier.

This complexity should go away when we upgrade to Clang 13 and later across the board.