[skip github actions] Apply Graviton-specific flags to both C and C++ flags
Prior to this commit, on aarch64, we were only applying special flags (including disabling the use of outline atomics) to C compiler flags but not to C++ compiler flags. Applying those flags to C++ as well. This is fixing some undefined symbols that occur during YugabyteDB build with Clang 15 on aarch64.

Skipping GitHub Actions because we currently only build for x86_64 on GitHub Actions.