Compile opentelemetery libraries for distributed tracing of YSQL queries (#191)
This commit compiles the latest (v1.9.0) version of opentelemetry-cpp. This version uses a protobuf feature experimental_allow_proto3_optional that is not supported by the protobuf version (v3.5.1) used in Yugabyte. Therefore a couple of patches to remove the usage of the feature has also been added. Another patch to fix undefined symbols error on macOS was added.