Support building DiskANN with GCC on x86_64 Linux (#270)
Support building DiskANN with GCC 11+ on x86_64 Linux by supplying the
correct OpenMP compiler flag.

Build a recent version of the patchelf tool so that it would recognize
new ELF features generated by recent compilers. This is only done with
GCC compilers because in the Clang build we use patchelf to patch
libraries including libc++ that the custom-built patchelf itself ends up
using. Adding a new build group CXX_UNINSTRUMENTED specifically for
patchelf, because it is a C++ dependency that does not need to be
ASAN/UBSAN/TSAN instrumented.

Use the `-Wl,--enable-new-dtags` linker flag so that the newer RUNPATH
attribute would be used instead of RPATH in the generated executables
and libraries.

Updating DiskANN to a patched version that explicitly specifies libaio
as a dependency of libdiskann.so, and getting rid of the custom `-laio`
linker flag specified externally.