[#21963] YSQL: Add build definition for AWS clockbound FFI. (#280)
https://github.com/aws/clock-bound lets us retrieve clock accuracies.
This is accurate as long as the reference clock is accurate. This
information helps us improve transaction performance.

Compiling clock-bound-ffi using rust generates a C library for use in
YugabyteDB. This source code is licensed Apache 2.0.

As a part of this change, build clock-bound-ffi using rust. We package
only the header file and the static library (not the shared library.
Using static library consistently across all the build types reduces
surprises).

clock-bound-ffi is a C library and not built with sanitizer support.

The option `--toolchain=llvm17` has no bearing in how clock-bound-ffi is
built.

---

Also, replace ubuntu 23.04 with ubuntu 24.04 since (1) 24.04 is the
newer and stable version and (2) yugabyte's 23.04 docker image does not
have rust installed.