[DEVOPS-3278] Move thirdparty builds off of amzn2 (#294)
While amzn2 does not go EOL until mid-2025, it is no longer supported on
YBDB releases beyond 2.20. YBDB builds on master branch build on later
OS, so are not compatible with older glibc.

Thirdparty dependency builds on amzn2 are still supported on the 2.20
branch of this repo.

Also fix python build issue on macos.