Disable alignment for bson types (#314)
PG15 which is what yb is currently based off of does not support aligned
memory allocations. The DocumentDB extension also uses non aligned
memory for storing the bson type in a few places. When the type is
natively used in DocDB there will also be no guarantee that the memory
is aligned. Hence disabling alignment.